chore: move tests to util_test
